python_ml_service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
requests==2.31.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.2.0
torchaudio==2.2.0
transformers==4.40.0
huggingface-hub==0.19.4
whisper==1.1.10
langchain==0.0.206
openai==0.28.0
python-docx==0.8.11
pdfminer.six==20221105
nltk==3.8.1
tqdm==4.66.1
pyyaml==6.0
python-dotenv